Authorization for edit and delete

+++ app/Http/Controllers/ListingController.php
@@ -46,10 +46,14 @@ class ListingController extends Controller
 
     public function edit(Listing $listing) {
 //      dd($listing->title);
+      if($listing->user_id != auth()->user()->id) return redirect('/')->with('message', 'Access Denied');
       return view('listings.edit')->with('listing', $listing);
     }
 
   public function update(Request $request, Listing $listing) {
+      if($listing->user_id != auth()->id()) {
+        abort(403, 'Unauthorized Action');
+      }
 //      dd($request->file('logo'));
     $formFields = $request->validate([
       'title' => 'required',
@@ -68,7 +72,15 @@ class ListingController extends Controller
   }
 
   public function destroy(Listing $listing) {
+    if($listing->user_id != auth()->id()) {
+      abort(403, 'Unauthorized Action');
+    }
+
     $listing->delete();
     return redirect('/')->with('message', 'Listing deleted successfully!');
+  }
+
+  public function manage() {
+      return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
   }
 }
